@@ -1,21 +1,30 @@
 # Copyright 2025 © BeeAI a Series of LF Projects, LLC
 # SPDX-License-Identifier: Apache-2.0
 
-from enum import Enum
+from enum import StrEnum
 
-from pydantic import AnyUrl, BaseModel
+from a2a.types import Role
+from pydantic import BaseModel, computed_field
+
+from beeai_sdk.platform import File
 
 
-class OriginType(str, Enum):
+class OriginType(StrEnum):
     UPLOADED = "uploaded"
     GENERATED = "generated"
 
 
+ORIGIN_TYPE_BY_ROLE = {
+    Role.user: OriginType.UPLOADED,
+    Role.agent: OriginType.GENERATED,
+}
+
+
 class FileChatInfo(BaseModel):
-    id: str
-    url: AnyUrl
-    filename: str
+    file: File
     display_filename: str  # A sanitized version of the filename used for display, in case of naming conflicts.
-    content_type: str | None = None
-    file_size_bytes: int | None = None
-    origin_type: OriginType
+    role: Role
+
+    @computed_field
+    def origin_type(self) -> OriginType:
+        return ORIGIN_TYPE_BY_ROLE[self.role]
