@@ -1,37 +1,33 @@
 # Copyright 2025 © BeeAI a Series of LF Projects, LLC
 # SPDX-License-Identifier: Apache-2.0
-
-import os
+import asyncio
 import re
+from contextlib import suppress
 from typing import Iterable
 
+import pydantic
 from a2a.types import FilePart, FileWithUri, Message, Role
 from beeai_framework.backend import AssistantMessage, UserMessage
 
-from chat.helpers.platform import get_file_info
-from chat.tools.files.model import FileChatInfo, OriginType
+from beeai_sdk.platform import File
+from beeai_sdk.util.file import PlatformFileUrl
+from chat.tools.files.model import FileChatInfo
 
 FrameworkMessage = UserMessage | AssistantMessage
 
 
 def to_framework_message(message: Message) -> FrameworkMessage:
-    message_text = "".join(
-        part.root.text for part in message.parts if part.root.kind == "text"
-    )
-
-    if message.role == Role.agent:
-        return AssistantMessage(message_text)
-
-    if message.role == Role.user:
-        return UserMessage(message_text)
-
-    raise ValueError(f"Invalid message role: {message.role}")
+    message_text = "".join(part.root.text for part in message.parts if part.root.kind == "text")
+    match message.role:
+        case Role.agent:
+            return AssistantMessage(message_text)
+        case Role.user:
+            return UserMessage(message_text)
+        case _:
+            raise ValueError(f"Invalid message role: {message.role}")
 
 
-async def extract_files(
-    history: list[Message],
-    incoming_message: Message,
-) -> list[FileChatInfo]:
+async def extract_files(history: list[Message], incoming_message: Message) -> list[FileChatInfo]:
     """
     Extracts file URLs from the chat history and the current turn's messages.
 
@@ -40,77 +36,39 @@ async def extract_files(
         incoming_message (Message): The message from the current turn.
 
     Returns:
-        list[FileInfo]: A list of FileInfo objects containing file details.
+        list[FileChatInfo]: A list of FileInfo objects containing file details.
     """
     # 1. Combine historical messages with the current turn
     all_messages = [*history, incoming_message]
 
-    # 2. Flatten to all parts
-    all_parts = (part.root for message in all_messages for part in message.parts)
-
-    # 3. Collect, validate, deduplicate while preserving order
+    # 2. Collect, validate, deduplicate while preserving order
     seen: set[str] = set()
-    files: list[FileChatInfo] = []
-    existing_filenames: list[str] = []
+    files: dict[str, Role] = {}
 
-    for part in all_parts:
-        if not isinstance(part, FilePart):
-            continue
+    for message in all_messages:
+        for part in message.parts:
+            match part.root:
+                case FilePart(file=FileWithUri(uri=uri)):
+                    with suppress(ValueError):
+                        url = pydantic.type_adapter.TypeAdapter(PlatformFileUrl).validate_python(uri)
+                        if url.file_id not in seen:
+                            seen.add(url.file_id)
+                            files[url.file_id] = message.role
 
-        if not isinstance(part.file, FileWithUri):
-            continue
-
-        url = part.file.uri.replace(
-            "{platform_url}", os.getenv("PLATFORM_URL", "127.0.0.1:8333")
-        )
-        if url and url not in seen:
-            fileInfo = await get_file_chat_info(
-                url,
-                content_type=part.file.mimeType or "application/octet-stream",
-                existing_filenames=existing_filenames,
-                origin_type=OriginType.UPLOADED,  # Default to uploaded since we can't access part.role
+    # TODO: N+1 query issue, add bulk endpoint
+    file_objects = await asyncio.gather(*(File.get(file_id) for file_id in files))
+    existing_filenames = set()
+    result = []
+    for file, role in zip(file_objects, files.values()):
+        result.append(
+            FileChatInfo(
+                file=file,
+                role=role,
+                display_filename=next_unused_filename(file.filename, existing_filenames),
             )
-            if fileInfo:
-                files.append(fileInfo)
-                existing_filenames.append(fileInfo.display_filename)
-            seen.add(url)
-
-    return files
-
-
-async def get_file_chat_info(
-    url: str, content_type: str, origin_type: OriginType, existing_filenames: list[str]
-) -> FileChatInfo:
-    # 1. Extract the file-ID from the signed-download URL
-    file_id_match = re.search(r"/([^/]+)/content", str(url))
-    if not file_id_match:
-        raise ValueError(f"Could not extract file_id from: {url!r}")
-    file_id = file_id_match.group(1)
-
-    # 2. Ask the platform for the file metadata
-    file_response = await get_file_info(file_id)
-    if not file_response:
-        raise ValueError(f"File with ID {file_id} not found on the platform.")
-
-    # 3. Merge the API data with the known url & content_type
-    merged_payload = {
-        **file_response.model_dump(),  # → id, filename, file_size_bytes …
-        "url": url,  # override / add
-        "content_type": content_type,
-        "display_filename": next_unused_filename(
-            file_response.filename, existing_filenames
-        ),  # ensure unique display name
-        "origin_type": origin_type,
-    }
-
-    # 4. Validate & coerce with Pydantic
-    file_info = FileChatInfo.model_validate(merged_payload)
-
-    # 5. Extra safeguard: make sure filename is a string
-    if not isinstance(file_info.filename, str):
-        file_info.filename = str(file_info.filename)
-
-    return file_info
+        )
+        existing_filenames.add(result[-1].display_filename)
+    return result
 
 
 def next_unused_filename(name: str, existing: Iterable[str]) -> str:
