@@ -1,22 +1,20 @@
 # Copyright 2025 © BeeAI a Series of LF Projects, LLC
 # SPDX-License-Identifier: Apache-2.0
+import asyncio
+from contextlib import suppress
 
-import os
-import re
-from typing import Iterable
-
+import pydantic.type_adapter
 from a2a.types import FilePart, FileWithUri, Message, Role
 from beeai_framework.backend import AssistantMessage, UserMessage
 
 from beeai_sdk.platform import File
+from beeai_sdk.util.file import PlatformFileUrl
 
 FrameworkMessage = UserMessage | AssistantMessage
 
 
 def to_framework_message(message: Message) -> FrameworkMessage:
-    message_text = "".join(
-        part.root.text for part in message.parts if part.root.kind == "text"
-    )
+    message_text = "".join(part.root.text for part in message.parts if part.root.kind == "text")
 
     if message.role == Role.agent:
         return AssistantMessage(message_text)
@@ -27,10 +25,7 @@ def to_framework_message(message: Message) -> FrameworkMessage:
     raise ValueError(f"Invalid message role: {message.role}")
 
 
-async def extract_files(
-    history: list[Message],
-    incoming_message: Message,
-) -> list[File]:
+async def extract_files(history: list[Message], incoming_message: Message) -> list[File]:
     """
     Extracts file URLs from the chat history and the current turn's messages.
 
@@ -49,36 +44,20 @@ async def extract_files(
 
     # 3. Collect, validate, deduplicate while preserving order
     seen: set[str] = set()
-    files: list[File] = []
+    file_ids: list[str] = []
 
     for part in all_parts:
-        if not isinstance(part, FilePart):
-            continue
+        match part:
+            case FilePart(file=FileWithUri(uri=uri)):
+                with suppress(ValueError):
+                    url = pydantic.type_adapter.TypeAdapter(PlatformFileUrl).validate_python(uri)
+                    if url.file_id not in seen:
+                        seen.add(url.file_id)
+                        file_ids.append(url.file_id)
 
-        if not isinstance(part.file, FileWithUri):
-            continue
+    # TODO: N+1 query issue, add bulk endpoint
+    return await asyncio.gather(*(File.get(file_id) for file_id in file_ids))
 
-        url = part.file.uri.replace(
-            "{platform_url}", os.getenv("PLATFORM_URL", "127.0.0.1:8333")
-        )
-        # if url and url not in seen:
-        #     fileInfo = await get_file_chat_info(
-        #         url,
-        #         content_type=part.file.mime_type or "application/octet-stream",
-        #         existing_filenames=existing_filenames,
-        #         origin_type=OriginType.UPLOADED,  # Default to uploaded since we can't access part.role
-        #     )
-        #     if fileInfo:
-        #         files.append(fileInfo)
-        #         existing_filenames.append(fileInfo.display_filename)
-        #     seen.add(url)
-        if url and url not in seen:
-            file = await File.get(extract_file_id_from_url(url))
-            if file:
-                files.append(file)
-            seen.add(url)
-
-    return files
 
 def format_size(size: int | None) -> str:
     if size is None:
@@ -89,23 +68,3 @@ def format_size(size: int | None) -> str:
         return f"{size / 1024:.1f} KB"
     else:
         return f"{size / (1024 * 1024):.1f} MB"
-
-
-def extract_file_id_from_url(url: str) -> str:
-    """
-    Extract file ID from platform file URL format.
-
-    Expected format: http://{platform_url}/api/v1/files/{file_id}/content
-    Args:
-        url (str): The URL from which to extract the file ID.
-    Raises:
-        ValueError: If the file ID cannot be extracted from the URL
-
-    Examples:
-        >>> extract_file_id_from_url("http://localhost:8333/api/v1/files/51797992-1dab-4c4d-bdac-9e90c28bbef1/content")
-        '51797992-1dab-4c4d-bdac-9e90c28bbef1'
-    """
-    file_id_match = re.search(r"/files/([^/]+)/content", url)
-    if not file_id_match:
-        raise ValueError(f"Could not extract file_id from URL: {url}")
-    return file_id_match.group(1)
