@@ -3,7 +3,7 @@
 
 import inspect
 from collections.abc import AsyncIterator, Callable
-from contextlib import asynccontextmanager
+from contextlib import AbstractAsyncContextManager, asynccontextmanager
 from inspect import isclass
 from typing import Annotated, Any, Generic, get_args, get_origin
 
@@ -38,16 +38,18 @@ class Depends(Generic[ExtensionSpecT, MetadataFromClientT]):
         if isinstance(dependency, BaseExtensionServer):
             self.extension = dependency
 
-    def __call__(self, message: Message, context: RunContext) -> Any:
-        return self._dependency_callable(message, context)
+    def __call__(self, message: Message, context: RunContext) -> AbstractAsyncContextManager[Any]:
+        instance = self._dependency_callable(message, context)
 
-    @asynccontextmanager
-    async def lifespan(self) -> AsyncIterator[None]:
-        if self.extension:
-            async with self.extension.lifespan():
-                yield
-        else:
-            yield
+        @asynccontextmanager
+        async def lifespan() -> AsyncIterator[Any]:
+            if self.extension or hasattr(instance, "lifespan"):
+                async with instance.lifespan():
+                    yield instance
+            else:
+                yield instance
+
+        return lifespan()
 
 
 def extract_dependencies(sign: inspect.Signature) -> dict[str, Depends]:
