@@ -1,6 +1,8 @@
 # Copyright 2025 © BeeAI a Series of LF Projects, LLC
 # SPDX-License-Identifier: Apache-2.0
-from typing import Annotated, Literal
+import base64
+import struct
+import typing
 
 import fastapi
 import ibm_watsonx_ai
@@ -10,6 +12,7 @@ import openai.types
 import pydantic
 from fastapi import Depends
 from fastapi.concurrency import run_in_threadpool
+from openai.types.create_embedding_response import CreateEmbeddingResponse
 
 from beeai_server.api.dependencies import EnvServiceDependency, RequiresPermissions
 from beeai_server.domain.models.permissions import AuthorizedUser
@@ -26,14 +29,22 @@ class EmbeddingsRequest(pydantic.BaseModel):
 
     model: str
     input: list[str] | str
-    encoding_format: Literal["float"] | None = None
+    encoding_format: typing.Literal["float", "base64"] | None = None
+
+
+class MultiformatEmbedding(openai.types.Embedding):
+    embedding: str | list[float]
+
+
+def _float_list_to_base64(embedding: list[float]) -> str:
+    return base64.b64encode(struct.pack(f"<{len(embedding)}f", *embedding)).decode("utf-8")
 
 
 @router.post("/embeddings")
 async def create_embedding(
     env_service: EnvServiceDependency,
     request: EmbeddingsRequest,
-    _: Annotated[AuthorizedUser, Depends(RequiresPermissions(embeddings={"*"}))],
+    _: typing.Annotated[AuthorizedUser, Depends(RequiresPermissions(embeddings={"*"}))],
 ):
     env = await env_service.list_env()
     backend_url = pydantic.HttpUrl(env["EMBEDDING_API_BASE"])
@@ -57,10 +68,14 @@ async def create_embedding(
             object="list",
             model=watsonx_response["model_id"],
             data=[
-                openai.types.Embedding(
+                MultiformatEmbedding(
                     object="embedding",
                     index=i,
-                    embedding=result["embedding"],
+                    embedding=(
+                        _float_list_to_base64(result["embedding"])
+                        if request.encoding_format == "base64"
+                        else typing.cast(list[float], result["embedding"])
+                    ),
                 )
                 for i, result in enumerate(watsonx_response.get("results", []))
             ],
@@ -70,16 +85,24 @@ async def create_embedding(
             ),
         ).model_dump(mode="json") | {"beeai_proxy_version": BEEAI_PROXY_VERSION}
     else:
-        return (
-            await openai.AsyncOpenAI(
-                api_key=env["EMBEDDING_API_KEY"],
-                base_url=str(backend_url),
-                default_headers=(
-                    {"RITS_API_KEY": env["EMBEDDING_API_KEY"]}
-                    if backend_url.host.endswith(".rits.fmaas.res.ibm.com")
-                    else {}
-                ),
-            ).embeddings.create(
-                **(request.model_dump(mode="json", exclude_none=True) | {"model": env["EMBEDDING_MODEL"]})
-            )
-        ).model_dump(mode="json") | {"beeai_proxy_version": BEEAI_PROXY_VERSION}
+        result: CreateEmbeddingResponse = await openai.AsyncOpenAI(
+            api_key=env["EMBEDDING_API_KEY"],
+            base_url=str(backend_url),
+            default_headers=(
+                {"RITS_API_KEY": env["EMBEDDING_API_KEY"]}
+                if backend_url.host.endswith(".rits.fmaas.res.ibm.com")
+                else {}
+            ),
+        ).embeddings.create(**(request.model_dump(mode="json", exclude_none=True) | {"model": env["EMBEDDING_MODEL"]}))
+        # Despite the typing, OpenAI library does return str embeddings when base64 is requested
+        # However, some providers, like Ollama, silently don't support base64, so we have to convert
+        if request.encoding_format == "base64" and result.data and isinstance(result.data[0].embedding, list):
+            result.data = [
+                MultiformatEmbedding(
+                    object="embedding",
+                    index=embedding.index,
+                    embedding=_float_list_to_base64(embedding.embedding),
+                )
+                for embedding in result.data
+            ]
+        return result.model_dump(mode="json") | {"beeai_proxy_version": BEEAI_PROXY_VERSION}
