@@ -6,23 +6,23 @@
 import type { NavItem } from '#components/SidePanel/Nav.tsx';
 import { Nav } from '#components/SidePanel/Nav.tsx';
 import { useRouteTransition } from '#contexts/TransitionContext/index.ts';
-import { useAgentNameFromPath } from '#hooks/useAgentNameFromPath.ts';
+import { useProviderIdFromUrl } from '#hooks/useProviderIdFromUrl.ts';
 import { useListAgents } from '#modules/agents/api/queries/useListAgents.ts';
 import { routes } from '#utils/router.ts';
 
 export function AgentsNav() {
-  const agentName = useAgentNameFromPath();
+  const providerId = useProviderIdFromUrl();
   const { transitionTo } = useRouteTransition();
 
   const { data: agents } = useListAgents({ onlyUiSupported: true, sort: true });
 
-  const items: NavItem[] | undefined = agents?.map(({ name }) => {
-    const route = routes.agentRun({ name });
+  const items: NavItem[] | undefined = agents?.map(({ name, provider: { id } }) => {
+    const route = routes.agentRun({ providerId: id });
 
     return {
-      key: name,
+      key: id,
       label: name,
-      isActive: agentName === name,
+      isActive: providerId === id,
       onClick: () => transitionTo(route),
     };
   });
