@@ -12,7 +12,7 @@ import { MarkdownContent } from '#components/MarkdownContent/MarkdownContent.tsx
 import { SidePanel } from '#components/SidePanel/SidePanel.tsx';
 import { useApp } from '#contexts/App/index.ts';
 import { SidePanelVariant } from '#contexts/App/types.ts';
-import { useAgentNameFromPath } from '#hooks/useAgentNameFromPath.ts';
+import { useProviderIdFromUrl } from '#hooks/useProviderIdFromUrl.ts';
 
 import { useAgent } from '../api/queries/useAgent';
 import { AgentCredits } from './AgentCredits';
@@ -21,8 +21,8 @@ import { AgentTags } from './AgentTags';
 import { AgentTools } from './AgentTools';
 
 export function AgentDetailPanel() {
-  const agentName = useAgentNameFromPath();
-  const { data: agent, isPending } = useAgent({ name: agentName ?? '' });
+  const providerId = useProviderIdFromUrl();
+  const { data: agent, isPending } = useAgent({ providerId: providerId ?? '' });
   const { activeSidePanel } = useApp();
 
   if (!agent) return null;
