@@ -11,6 +11,7 @@ import { TextAreaAutoHeight } from '#components/TextAreaAutoHeight/TextAreaAutoH
 import { InteractionMode } from '#modules/agents/api/types.ts';
 import { FileUploadButton } from '#modules/files/components/FileUploadButton.tsx';
 import { useFileUpload } from '#modules/files/contexts/index.ts';
+import { usePlatformContext } from '#modules/platform-context/contexts/index.ts';
 import { dispatchInputEventOnTextarea, submitFormOnEnter } from '#utils/form-utils.ts';
 
 import { ChatDefaultTools } from '../chat/constants';
@@ -32,6 +33,8 @@ export function RunInput({ promptExamples, onSubmit }: Props) {
   const inputRef = useRef<HTMLTextAreaElement>(null);
 
   const [promptExamplesOpen, setPromptExamplesOpen] = useState(false);
+
+  const { contextId } = usePlatformContext();
 
   const {
     agent: {
@@ -56,7 +59,7 @@ export function RunInput({ promptExamples, onSubmit }: Props) {
 
   const inputProps = register('input', { required: true });
   const inputValue = watch('input');
-  const isSubmitDisabled = isPending || isFileUploadPending || !inputValue;
+  const isSubmitDisabled = isPending || isFileUploadPending || !inputValue || !contextId;
 
   const dispatchInputEventAndFocus = useCallback(() => {
     const inputElem = inputRef.current;
