@@ -7,7 +7,7 @@ export const routes = {
   home: () => '/' as const,
   login: () => '/signin' as const,
   notFound: () => '/not-found' as const,
-  agentRun: ({ name }: { name: string }) => `/run?agent=${name}`,
+  agentRun: ({ providerId }: { providerId: string }) => `/run?p=${providerId}`,
   playground: () => '/playground' as const,
   playgroundSequential: () => '/playground/sequential' as const,
   settings: () => '/settings' as const,
